+++ code/engine.vc2008/xrRender/xrRender/WallmarksEngine.cpp
@@ -206,7 +206,7 @@ void CWallmarksEngine::AddWallmark_internal	(CDB::TRI* pTri, const Fvector* pVer
 		bb_query.get_CD		(bbc,bbd);
 		xrc.box_options		(CDB::OPT_FULL_TEST);
 		xrc.box_query		(g_pGameLevel->ObjectSpace.GetStaticModel(),bbc,bbd);
-		u32	triCount		= xrc.r_count	();
+		u32	triCount		= (u32)xrc.r_count	();
 		if (0==triCount)	
 			return;
